Replaced uint with IntPtr for handles.

+++ Upnp/Public/Cs/CpDeviceUpnp.cs
@@ -6,7 +6,7 @@ namespace Zapp
     public class CpDeviceListUpnpAll : CpDeviceList
     {
         [DllImport ("ZappUpnp")]
-        static extern unsafe uint CpDeviceListCreateUpnpAll(CallbackDevice aAdded, IntPtr aPtrAdded,
+        static extern unsafe IntPtr CpDeviceListCreateUpnpAll(CallbackDevice aAdded, IntPtr aPtrAdded,
                                                             CallbackDevice aRemoved, IntPtr aPtrRemoved);
 
         public unsafe CpDeviceListUpnpAll(ChangeHandler aAdded, ChangeHandler aRemoved)
@@ -21,7 +21,7 @@ namespace Zapp
     public class CpDeviceListUpnpRoot : CpDeviceList
     {
         [DllImport ("ZappUpnp")]
-        static extern unsafe uint CpDeviceListCreateUpnpRoot(CallbackDevice aAdded, IntPtr aPtrAdded,
+        static extern unsafe IntPtr CpDeviceListCreateUpnpRoot(CallbackDevice aAdded, IntPtr aPtrAdded,
                                                              CallbackDevice aRemoved, IntPtr aPtrRemoved);
 
         public unsafe CpDeviceListUpnpRoot(ChangeHandler aAdded, ChangeHandler aRemoved)
@@ -36,7 +36,7 @@ namespace Zapp
     public class CpDeviceListUpnpUuid : CpDeviceList
     {
         [DllImport ("ZappUpnp", CharSet = CharSet.Ansi)]
-        static extern unsafe uint CpDeviceListCreateUpnpUuid(char* aUuid,
+        static extern unsafe IntPtr CpDeviceListCreateUpnpUuid(char* aUuid,
                                                              CallbackDevice aAdded, IntPtr aPtrAdded,
                                                              CallbackDevice aRemoved, IntPtr aPtrRemoved);
 
@@ -54,7 +54,7 @@ namespace Zapp
     public class CpDeviceListUpnpDeviceType : CpDeviceList
     {
         [DllImport ("ZappUpnp", CharSet = CharSet.Ansi)]
-        static extern unsafe uint CpDeviceListCreateUpnpDeviceType(char* aDomainName, char* aDeviceType, uint aVersion,
+        static extern unsafe IntPtr CpDeviceListCreateUpnpDeviceType(char* aDomainName, char* aDeviceType, uint aVersion,
                                                                    CallbackDevice aAdded, IntPtr aPtrAdded,
                                                                    CallbackDevice aRemoved, IntPtr aPtrRemoved);
 
@@ -75,7 +75,7 @@ namespace Zapp
     public class CpDeviceListUpnpServiceType : CpDeviceList
     {
         [DllImport ("ZappUpnp", CharSet = CharSet.Ansi)]
-        static extern unsafe uint CpDeviceListCreateUpnpServiceType(char* aDomainName, char* aServiceType, uint aVersion,
+        static extern unsafe IntPtr CpDeviceListCreateUpnpServiceType(char* aDomainName, char* aServiceType, uint aVersion,
                                                                     CallbackDevice aAdded, IntPtr aPtrAdded,
                                                                     CallbackDevice aRemoved, IntPtr aPtrRemoved);
